KAN-96 feat: refactor
