+++ src/services/incragram.base-query.ts
@@ -1,6 +1,7 @@
 import { getToken, setToken } from '@/shared/utils/storage'
 import { BaseQueryFn, FetchArgs, FetchBaseQueryError, fetchBaseQuery } from '@reduxjs/toolkit/query'
 import { Mutex } from 'async-mutex'
+import NProgress from 'nprogress'
 
 const mutex = new Mutex()
 const baseQuery = fetchBaseQuery({
@@ -21,6 +22,7 @@ export const baseQueryWithReauth: BaseQueryFn<
   unknown,
   FetchBaseQueryError
 > = async (args, api, extraOptions) => {
+  NProgress.start()
   await mutex.waitForUnlock()
   let result = await baseQuery(args, api, extraOptions)
 
@@ -57,6 +59,7 @@ export const baseQueryWithReauth: BaseQueryFn<
       result = await baseQuery(args, api, extraOptions)
     }
   }
+  NProgress.done()
 
   return result
 }
