+++ src/shared/types/public.enums.ts
@@ -1,0 +1,11 @@
+export enum BrowserList {
+  Brave = 'brave',
+  Chrome = 'chrome',
+  Edge = 'edge',
+  Explorer = 'explorer',
+  Firefox = 'firefox',
+  Opera = 'opera',
+  Safari = 'safari',
+  Uc = 'uc',
+  Yandex = 'yandex',
+}
